requirements.txt
Pillow>=9.0.0
emoji>=2.0.0
zhdate>=0.1.0